+++ source/view/player/player.cpp
@@ -5,7 +5,7 @@
 constexpr auto SPRITE_HEIGHT = 120.0f;
 constexpr auto SPRITE_WIDTH = 10.0f;
 
-Player::Player(const sf::Vector2f initial_position, const float max_height)
+Player::Player(const sf::Vector2f initial_position, const double max_height)
 	: m_position(initial_position), m_max_height(max_height),
 	  m_rect_shape(sf::Vector2f(SPRITE_WIDTH, SPRITE_HEIGHT)) {
 
@@ -14,12 +14,13 @@ Player::Player(const sf::Vector2f initial_position, const float max_height)
 
 Player::~Player() {}
 
-auto Player::move(float delta_y) -> void {
+auto Player::move(double delta_y) -> void {
 	auto& [x, y] = m_position;
 	y += delta_y;
 
 	// Clamp it to the max height of the screen - that of the player
-	y = std::clamp(y, 0.0f, this->m_max_height - SPRITE_HEIGHT);
+	y = std::clamp(y, 0.0f,
+				   static_cast<float>(this->m_max_height) - SPRITE_HEIGHT);
 
 	this->m_rect_shape.setPosition({x, y});
 }
